nxxlsx_loadgs.py
import pandas as pd
from googleapiclient.discovery import build
from google_auth_oauthlib.flow import InstalledAppFlow,Flow
from google.auth.transport.requests import Request
import os
import pickle

SCOPES = ['https://www.googleapis.com/auth/spreadsheets']

# here enter the id of your google sheet
SPREADSHEET_ID_input='1TwcW3o3zUz2uYpG2on1-LSZCSb3P7UDtZZV4XuzsVuk'
RANGE_NAME = 'A:J'

def main():
    global values_input, service
    creds = None
    if os.path.exists('token.pickle'):
        with open('token.pickle', 'rb') as token:
            creds = pickle.load(token)
    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(
                'credentials.json', SCOPES) # here enter the name of your credentials JSON file
            creds = flow.run_local_server(port=0)
        with open('token.pickle', 'wb') as token:
            pickle.dump(creds, token)

    service = build('sheets', 'v4', credentials=creds)

    # Call the Sheets API
    sheet = service.spreadsheets()
    result_input = sheet.values().get(spreadsheetId=SPREADSHEET_ID_input,
                                range=RANGE_NAME).execute()
    values_input = result_input.get('values', [])

    if not values_input and not values_expansion:
        print('No data found.')

main()

df=pd.DataFrame(values_input[1:], columns=values_input[0])

# Writing the data into the excel sheet
writer_obj = pd.ExcelWriter('maListeNetflix.xlsx',
                            engine='xlsxwriter')
df.to_excel(writer_obj, sheet_name='Liste')
writer_obj.save()
print('INFO - nxxlsx_loadgs.py - Le fichier maListeNetflix.xlsx est chargé')