requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
future==0.18.2
google-api-core==1.25.1
google-api-python-client==1.12.8
googleapis-common-protos==1.52.0
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-spreadsheet==0.0.6
html5lib==1.1
httplib2==0.18.1
idna==2.10
Jinja2==2.11.3
joblib==1.0.1
livereload==2.6.3
lunr==0.5.8
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==7.0.6
mkdocs-material-extensions==1.0.1
nltk==3.5
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pymdown-extensions==8.1.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
selenium==3.141.0
six==1.15.0
soupsieve==2.1
tornado==6.1
tqdm==4.59.0
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
wkhtmltopdf
XlsxWriter==1.3.7
